# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/logs/error/{lines}": {
      "get": {
        "operationId": "LogsController_getErrorLogs",
        "parameters": [
          {
            "name": "lines",
            "required": true,
            "in": "path",
            "description": "Number of error log lines to retrieve",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "logs"
        ]
      }
    },
    "/logs/output/{lines}": {
      "get": {
        "operationId": "LogsController_getOutputLogs",
        "parameters": [
          {
            "name": "lines",
            "required": true,
            "in": "path",
            "description": "Number of output log lines to retrieve",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "logs"
        ]
      }
    },
...
  },
  "info": {
    "title": "",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "AdminKey": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "stack": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "name",
          "stack"
        ]
      },
...
      "LogResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDto"
          }
        },
        "required": [
          "success",
          "data",
          "error"
        ]
      }
    }
  }
}