# web/api/models/ErrorDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ErrorDto = {
    message: string;
    name: string;
    stack: string;
};